src/main/java/com/wj/resttasks/service/UserServiceDetailsImpl.java
package com.wj.resttasks.service;

import com.wj.resttasks.models.User;
import com.wj.resttasks.repositories.UserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.core.GrantedAuthority;
import org.springframework.security.core.authority.SimpleGrantedAuthority;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

import java.util.ArrayList;
import java.util.List;
import java.util.Optional;

@Service
public class UserServiceDetailsImpl implements UserDetailsService {

    @Autowired
    UserRepository userRepository;

    @Override
    public UserDetails loadUserByUsername(String username)
        throws UsernameNotFoundException {
        Optional<User> user = userRepository.findByUsername(username);
        org.springframework.security.core.userdetails.User.UserBuilder builder = null;

        if (user != null) {

            User currentUser = user.get();

            //create an authority from User's role and add it to the UserDetails builder
            List<GrantedAuthority> grantedAuthorities = new ArrayList<>();
            grantedAuthorities.add(new SimpleGrantedAuthority("ROLE_" + currentUser.getRole()));

            builder = org.springframework.security.core.userdetails.User.withUsername(username);
            builder.password(currentUser.getPassword());
            builder.authorities(grantedAuthorities);
        } else {
            throw new UsernameNotFoundException("User not found");
        }
        return builder.build();
    }

    public UserServiceDetailsImpl(UserRepository userRepository) {
        this.userRepository = userRepository;
    }
}
